ans-tower windows: generate a password
When creating the windows machines, we change the Administrator password in the
Cloudformation template, using user-data and a powershell command. The current
password was hardcoded in the Cloudformation template.

We want to be able to set the windows password with the variable
'windows_password', either in env_secret_vars.yml or with
'-e windows_password=YOURPASS'. If no windows_variables is passed to
ansible_agnostic_deployer, the default behavior is to randomly generate one.

This commit gets rid of the hardcoded password:
- generate one if windows_password variable is not defined
- save the password into workdir/[...]_windows_password.txt for users.
- set the fact generated_windows_password when appropriate and use it later in
  the playbooks in case windows_password is not defined. Like this:
  {{ windows_password |
  default(hostvars['localhost'].generated_windows_password) }}

Also:
- in the ansible inventory file on bastion, use the same ansible_* variables
  for windows vm access: add ansible_ssh_user=Administrator and
  ansible_become=false