cloudformation templates: fix race condition
The problem during the stress tests appears to be a race condition established
between the creation of the Elastic IP resource and the VPC Gateway Attachment
resource that connects the created VPC to the wider internet. For the EIP to be
created successfully, it needs to be created in a VPC with an established
network connection.

It appears that in most cases, the Gateway Attachment is created before the
EIPs, and the stack is therefore successful. The failures occur when the EIP
creation is attempted before the Gateway Attachment. To avoid this condition,
this commit makes the EIP resources dependent on the Gateway Attachment using a
DependsOn statement.