ocp 3.7: fix elasticsearch and cassandra on NFS
2 scenarii here:

1. without glusterfs

This set of variable works for 3.7 to keep ES and cassandra on NFS

see https://bugzilla.redhat.com/show_bug.cgi?id=1538995

2. with glusterfs

The idea is not to put ES and cassandra on glusterfs (not supported), instead
try to keep them on NFS. Currently this does not work, PVC for both will still
use default storageClass. It requires manual steps on PVC after the installation.
To do things properly, a new class glusterblock should be created  by the
openshift-ansible playbook. This glusterblock storageclass then to be used for ES
and cassandra PVC. This is not yet possible, see:

https://bugzilla.redhat.com/show_bug.cgi?id=1534750