ocp-workshop: do not use NFS with glusterfs
If user decides to install glusterfs, then use default 'EmptyDir' for
logging (elasticsearch) and metrics (cassandra).

Why? Because we set glusterfs storageClass as default to be able to use
'dynamic' as storage type for some components (prometheus, ...). When there is a
default storageClass, openshift-ansible does not setup ES and Cassandra on NFS
even if told so.

User can still setup ES and Cassandra storage in post-install.

( NFS is going to be unsupported and removed completely in future versions. )