Add more dependencies in CloudFormation stacks
- create RecordSetGroup after instances
- ocp-workshop: in cloudformation create the S3 at the end

We we try to fallback to another region, it is sometimes stuck in
'CREATE_IN_PROGRESS' because of the s3bucket.
This will make the s3 bucket one of the last resources to be created.
This way, when we fallback to another region, there is a very good chance it's
not yet created.