detect and set AWS region in case aws_region_final is not defined
Sometimes the infra step is skipped, for example when scaling up a cluster.
when step001.1 is skipped, aws_region_final is not defined.

This commits adds tasks to detect the region and set aws_region_final depending
on 'fallback_regions' being defined or not.

Also update the destroy.yml play to use those tasks (DRY code).