Use default Cloudformation template in ocp4-coreos-deployer
This allows the creation of a dedicated subzone. This will facilitate the
installation of ocp4.

(idea) Also write down the Route53 User to the .aws/credentials file. This can be used
later when we know exactly what permission to grant the user in order for the
installation to work. That way the user belongs to the stack and gets deleted
when the stack is destroyed.