ocp4-workshop lifecycle: enable env_type when filtering instances (#793)
When stopping/starting instances of an environment, we are (temporarily)
filtering only by the tag 'guid'. It used to be filtered by ('guid',
'env_type'). It was changed to ('guid') because of a change in CF:
env_type was not passed as cloud_tags, and the instances only had guid
in their tags.

To be able to stop the environments we commented out tags with key 'env_type'.

Now this is fixed in Cloudforms, and also env_type and guid tags are
enforced in agnosticd.

All environments now have both 'guid' and 'env_type' tags.

This commit if applied will restore filters to ('guid', 'env_type').

(guid, env_type) is a better filter in case of guid collision.