ensure backwards compatibility ready for v1.6.x
in v1.6.x of integreatly we want to use a new inventory file, however the
older releases of integreatly should still maintain using the older hosts
inventory file.

this change adds a version check to ig_version, if it is within the current
known legacy versions that are installable from pds, the older hosts
inventory will be used. Else, the new pds inventory will be used.

The aim of this is to ensure the removal of this logic will be simple once
it is no longer needed.