Introduce default_vars.yml
`env_vars.yml` is a bad name and should be renamed to `default_vars.yml`.

This commit if applied, will update the `include_vars.yml` play so it looks
for the following files and includes them if they exist:

- `cloud_providers/{{ cloud_provider }}_default_vars.yml`
  Global default vars related to the cloud provider
- `configs/{{ env_type }}/env_vars.yml`
  Legacy env_vars.yml (replaced by default_vars.yml)
- `configs/{{ env_type }}/default_vars.yml`
  Default vars of the config
- `configs/{{ env_type }}/default_vars.{{ cloud_provider }}.yml`
  Default vars of the config, specific to a cloud provider
- `configs/{{ env_type }}/default_vars_{{ cloud_provider }}.yml`
  `configs/{{ env_type }}/env_secret_vars.yml`
  Lecacy secret vars file.

Both .yml and .yaml extensions are supported.

In order not to break current configs, `env_vars.yml` files are still included as
before.

This commit also makes it possible to have default vars for each cloud provider,
for example, it is now possible to have those files in a config:

```yaml
default_vars.yml
default_vars_osp.yml
default_vars_ec2.yml
```