poll_email.py: handle 'too many connections' error at login
This commit, if applied, adds the logic to retry when the IMAP login fails.

It implements the exponential backoff.

This commit fixes an error that occur when there are a significant number
of concurrent jenkins job running:

Traceback (most recent call last):
  File "./tests/jenkins/downstream/poll_email.py", line 111, in <module>
    args.timeout))
  File "./tests/jenkins/downstream/poll_email.py", line 65, in wait_email
    M = connect()
  File "./tests/jenkins/downstream/poll_email.py", line 54, in connect
    M.login(username, password)
  File "/usr/lib64/python2.7/imaplib.py", line 520, in login
    raise self.error(dat[-1])
imaplib.error: [ALERT] Too many simultaneous connections. (Failure)