Allow delete all resource when developing from a sandbox
Introduce osp_force_delete_all_resource variable.

If the user is using osp_project_create = false, he/she cannot delete the
resources inside the project.

If the user is alone in that project, he might want to just delete everything.

This commit introduces a new variable that allow that.

Also: prefer the osp_project_id when possible because that doesn't requirement
privileged access to list all the project.