ocp4-cluster: IOPS setting for master storage
Allow user to specify the desired IOPS for master storage.

In case of large clusters the master instance should have provisioned
IOPS for the etcd database to work properly.

This change, if applied, automatically sets the
`controlPlane.platform.aws.rootVolume.type` to `io1` if the number of
workers is greater than 10.

In that case, the IOPS value is calculated:

* minimum of 2000 IOPS,
* if number of wokers > 20, then it's 100 IOPS per node
* maximum of 32000 IOPS

According to OpenShift documentation

https://docs.openshift.com/container-platform/4.6/installing/installing_aws/installing-aws-customizations.html

* type = io1 / 2000 IOPS  is a good setting for large clusters