pass IP address of osp_cluster_dns_server to nsupdate
The `nsupdate` module expects the `server` to be an IP address, not a
hostname (see [1]), but osp_cluster_dns_server is a hostname.
Use the `dig` lookup to convert that to a real IP address.

If you pass a hostname to the `server` parameter, the Python DNS module
raises an error:

    dns.exception.SyntaxError: Text input is malformed.

[1] https://docs.ansible.com/ansible/latest/collections/community/general/nsupdate_module.html#parameter-server