Update config ocp4-cluster: Enable On-Demand Capacity Reservation (#3753)
* make agnosticd_odcr.py python2.7 compatible

* agnosticd_odcr: add distinct feature

add an option to force all the AZ to be distinct.

* Process reservation groups alphabetically

That makes ODCR creation predictable.

* Add example of complexe ODCR for testing

* Run odcr role in infra step only if it was not yet

That allows the role to be executed as pre_infra step in the config.

* enable ODCR for ocp4-cluster config

Define 2 ODCR profiles:

workshop:
- workers are split on 2 zones, if possible. Can be a single zone
  depending on availbility.
- Masters are all in the same zone.
- Bastion has its own zone, which can also be the same as the other zones
All zones can be a single zone depending on availability (not distinct)

ha: 3 distinct zones. Workers and masters are all on different zones. Higher risk of provision failures.

* Add py2 AWS docker file for testing ODCR plugin

* typo

* fix yamllint

* Address WK's and Nate's concern, simplify

* Remove example in just-a-bunch-of-nodes

* name block

* Remove outdated comment