Implement Open Environment checkbox feature (AWS)
This change, if applied, adds a new role to AgnosticD to
enable or disable Open Environment full access for any catalog item.

The variable is the boolean `agnosticd_open_environment`.
If `true`, the role `infra-aws-open-environment` is executed during the infra step.
It creates an admin user with programmatic and Web Console access.
It then prints the credentials to the user as user.info.

Also:

- create an infra-cloud-tags role to define in one place the
  default cloud tags and reuse them in several other roles or configs.