Add the possibility to use Ansible as student user
By default, only the root user can easily use Ansible throughout the environment, this is definitely not best practice, and having the student user being able to do it is better.
The solution chosen is a bit complex but avoids duplicating code.

Add possibility to define a group different from user name,
because the student user is created with the "users" group

Add flag to allow the student user to use Ansible or not,
variable is false by default to be backward compatible