Removing forced external subnet definition
This needs to be implemented as per the osp_find_public_subnet tag.
Currently, with the system forcing the public subnet, we will have large
numbers of errors whenever the next subnet is nearly full as there will
only be 1 or 2 FIP's available, and this will cause an IP cannot be
obtained error from the external network, even though there are other
free subnets.