Remove use_own_key option
This option push the user's key directly to the remote host. This is
wrong by design and it has been a point of confusion.

Plus, the feature is always disabled (we never push the key) by default.
I also checked in agnosticV, it's not used.

This change, if applied, removes entirely any use of that variable and
remove the tasks that push the private keys to the remote hosts entirely.