Allow symmetric encryption of output_dir tarball
Allow user to define a password to protect the output_dir in s3 in case
of a shared s3 bucket.

> agnosticd_save_output_dir_archive_password

Append '.gpg' extension in that case.

Use GNUPG tool to perform the symmetric encryption (by default it's
AES256), as it's the most available, even in images (it's more
available in images by default than openssl).