bastion-lite: Remove env_authorized_key dependency
This change, if applied, remove the dependency of env_authorized_key for
the bastion-lite role.

Instead, generate an SSH key on the bastion and use that to allow the
root@bastion user to ssh to any host.

That makes the bastion-lite role cloud-agnosticd.

- tested with just-a-bunch-of-nodes on equinix and AWS
- generate a key on the bastion and use that to setup additional access
  to all the hosts.
  The additional access is for the users when they ssh to the bastion:

```
ssh bastion...
sudo -i
ssh node1.GUID.internal
```